backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.5.0
pymongo==4.6.0
python-dotenv==1.0.0
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
aiohttp==3.9.1
email-validator==2.1.0
kaggle==1.6.12
pandas==1.5.3
reportlab==4.0.7
pypdf==4.0.1
dnspython==2.4.2